fix: address CodeRabbit review comments on PR #4031
- key-visual-form: reset selectedFile, dimensions, preview on validation error to avoid stale state.
- Split pure image validation (consts + validateImageFile) to new server-safe @ykzts/supabase/image-validation (no "use client"). Update server actions (upload-avatar, upload-key-visual) and client forms to import appropriately. Remove barrel re-export from image-upload.
- Fix extractTextFromPortableText to concat text within each block using empty separator (avoid intra-sentence spaces), join blocks with single space. Added explicit unknown typing for child to satisfy strict TS.
