+++ packages/supabase/src/image-upload.ts
@@ -1,6 +1,10 @@
 "use client";
 
 import { createBrowserClient } from "@ykzts/supabase/client";
+import {
+  MIME_TO_EXT,
+  validateImageFile,
+} from "@ykzts/supabase/image-validation";
 
 export interface ImageUploadOptions {
   file: File;
@@ -12,47 +16,6 @@ export interface ImageUploadResult {
   height?: number;
   url?: string;
   width?: number;
-}
-
-export const MAX_FILE_SIZE = 5 * 1024 * 1024; // 5MB
-export const ALLOWED_TYPES = [
-  "image/jpeg",
-  "image/png",
-  "image/gif",
-  "image/webp",
-];
-
-// Map MIME types to file extensions
-export const MIME_TO_EXT: Record<string, string> = {
-  "image/gif": "gif",
-  "image/jpeg": "jpg",
-  "image/png": "png",
-  "image/webp": "webp",
-};
-
-export function validateImageFile(file: File): { error?: string } | null {
-  if (!ALLOWED_TYPES.includes(file.type)) {
-    return {
-      error:
-        "サポートされていない画像形式です。JPEG、PNG、GIF、WebPのみアップロード可能です。",
-    };
-  }
-
-  if (file.size > MAX_FILE_SIZE) {
-    return {
-      error:
-        "ファイルサイズが大きすぎます。5MB以下の画像をアップロードしてください。",
-    };
-  }
-
-  const fileExt = MIME_TO_EXT[file.type];
-  if (!fileExt) {
-    return {
-      error: "不明な画像形式です。",
-    };
-  }
-
-  return null;
 }
 
 export async function uploadImage({
